+++ src/Core/EventListener.h
@@ -25,7 +25,7 @@ namespace crynn
 		/// Called once every frame.
 		/// </summary>
 		/// <param name="deltaTime">The time it took the previous frame to render.</param>
-		virtual void Update(double deltaTime) {}
+		virtual void Update(float deltaTime) {}
 		virtual void Start() {} ///Called once before the game enters its main loop. Will be called before the first Update.
 
 		//Called when the engine is going to render a frame
@@ -36,7 +36,7 @@ namespace crynn
 		/// Generally used for setting uniforms, or other things that need to be set before any render calls.
 		/// </summary>
 		/// <param name="deltaTime">The time it took the previous frame to render.</param>
-		virtual void BeforeUpdate(double deltaTime) {}
+		virtual void BeforeUpdate(float deltaTime) {}
 		virtual void BeforeClose() {} ///Called before quitting to desktop with Application::Instance().Quit();
 	private:		
 		///Subscribes the events in this class to the Crynn event system. Call at the end of your constructor.
@@ -45,11 +45,11 @@ namespace crynn
 		friend class CrynnObject;
 
 		//Used to remove handlers when this object is destructed.
-		unsigned int updateHandlerID = 0; 
-		unsigned int startHandlerID = 0;
-		unsigned int beforeUpdateHandlerID = 0;
-		unsigned int beforeCloseHandlerID = 0;
-		unsigned int renderID = 0;
+		uint64_t updateHandlerID = 0; 
+		uint64_t startHandlerID = 0;
+		uint64_t beforeUpdateHandlerID = 0;
+		uint64_t beforeCloseHandlerID = 0;
+		uint64_t renderID = 0;
 	};
 }
